+++ recipes/tests.py
@@ -6,18 +6,23 @@ from django.test import TestCase, Client
 from django.urls import reverse
 from django.utils.translation import activate
 
-from .models import Category, Unit, Tag, Ingredient, IngredientInRecipe, Recipe
+from .models import (
+    Category, Unit, Tag, Ingredient, IngredientUnit, IngredientInRecipe, Recipe
+)
 
 
 class RecipesTestCase(TestCase):
 
-    def ingredient_to_cart(self, ingredient, total):
+    def model_to_cart(self, ingredient_unit, total):
+        ingredient = ingredient_unit.ingredient
+        category = ingredient.category
+        unit = ingredient_unit.unit
         return [{
-            'ingredient__pk': ingredient.pk,
-            'ingredient__name': ingredient.name,
-            'ingredient__category__name': ingredient.category.name,
-            'ingredient__unit__pk': ingredient.unit.pk,
-            'ingredient__unit__name': ingredient.unit.name,
+            'ingredient_unit__pk': ingredient_unit.pk,
+            'ingredient_unit__ingredient__name': ingredient.name,
+            'ingredient_unit__ingredient__category__name': category.name,
+            'ingredient_unit__unit__pk': unit.pk,
+            'ingredient_unit__unit__name': unit.name,
         }, Decimal(total)]
 
     def setUp(self):
@@ -31,11 +36,15 @@ class RecipesTestCase(TestCase):
         self.ts, _ = Unit.objects.get_or_create(name='ts')
         self.g, _ = Unit.objects.get_or_create(name='g')
         self.ingredient1 = Ingredient.objects.create(
-            name='ingredient1', unit=self.pc, category=self.cat1)
-        self.ingredient2ts = Ingredient.objects.create(
-            name='ingredient2ts', unit=self.ts, category=self.cat2)
-        self.ingredient2g = Ingredient.objects.create(
-            name='ingredient2g', unit=self.g, category=self.cat2)
+            name='ingredient1', category=self.cat1)
+        self.ingredient1pc = IngredientUnit.objects.create(
+            ingredient=self.ingredient1, unit=self.pc)
+        self.ingredient2 = Ingredient.objects.create(
+            name='ingredient2', category=self.cat2)
+        self.ingredient2ts = IngredientUnit.objects.create(
+            ingredient=self.ingredient2, unit=self.ts)
+        self.ingredient2g = IngredientUnit.objects.create(
+            ingredient=self.ingredient2, unit=self.g)
         self.recipe1 = Recipe.objects.create(
             title='recipe1', recipe='Preparation recipe1')
         self.recipe2 = Recipe.objects.create(
@@ -44,16 +53,16 @@ class RecipesTestCase(TestCase):
         self.recipe1.tags.add(self.tag2)
         self.recipe2.tags.add(self.tag2)
         self.iir1_1 = IngredientInRecipe.objects.create(
-            recipe=self.recipe1, ingredient=self.ingredient1,
+            recipe=self.recipe1, ingredient_unit=self.ingredient1pc,
             amount=Decimal('3'), order=1)
         self.iir1_2ts = IngredientInRecipe.objects.create(
-            recipe=self.recipe1, ingredient=self.ingredient2ts,
+            recipe=self.recipe1, ingredient_unit=self.ingredient2ts,
             amount=Decimal('2'), order=2)
         self.iir2_1 = IngredientInRecipe.objects.create(
-            recipe=self.recipe2, ingredient=self.ingredient1,
+            recipe=self.recipe2, ingredient_unit=self.ingredient1pc,
             amount=Decimal('2'), order=1)
         self.iir2_2g = IngredientInRecipe.objects.create(
-            recipe=self.recipe2, ingredient=self.ingredient2g,
+            recipe=self.recipe2, ingredient_unit=self.ingredient2g,
             amount=Decimal('1'), order=2)
 
     def test_index(self):
@@ -92,7 +101,7 @@ class RecipesTestCase(TestCase):
         ctx = self.client.get(
             reverse('recipe', args=[self.recipe1.pk])).context
         # ingredient in recipe order
-        self.assertEqual(list(ctx['ingredients']),
+        self.assertEqual(list(ctx['ingredient_units']),
                          [self.iir1_2ts, self.iir1_1])
 
         ctx = self.client.get(reverse('cart')).context
@@ -100,17 +109,17 @@ class RecipesTestCase(TestCase):
         recipes = [(self.recipe2, Decimal(1)), (self.recipe1, Decimal(2))]
         self.assertEqual(ctx['recipes'], recipes)
         # ingredient cart order by category
-        ingredients = [self.ingredient_to_cart(self.ingredient2ts, 4),
-                       self.ingredient_to_cart(self.ingredient2g, 1),
-                       self.ingredient_to_cart(self.ingredient1, 8)]
-        self.assertEqual(ctx['ingredients'], ingredients)
+        ingredient_units = [self.model_to_cart(self.ingredient2ts, 4),
+                            self.model_to_cart(self.ingredient2g, 1),
+                            self.model_to_cart(self.ingredient1pc, 8)]
+        self.assertEqual(ctx['ingredient_units'], ingredient_units)
 
     def test_recipe(self):
         ctx = self.client.get(
             reverse('recipe', args=[self.recipe1.pk])).context
         self.assertEqual(ctx['page'], 'recipe')
         self.assertEqual(ctx['recipe'], self.recipe1)
-        self.assertEqual(list(ctx['ingredients']),
+        self.assertEqual(list(ctx['ingredient_units']),
                          [self.iir1_1, self.iir1_2ts])
 
     def test_cart(self):
@@ -122,10 +131,10 @@ class RecipesTestCase(TestCase):
         self.assertEqual(ctx['page'], 'cart')
         recipes = [(self.recipe1, Decimal(2)), (self.recipe2, Decimal(1))]
         self.assertEqual(ctx['recipes'], recipes)
-        ingredients = [self.ingredient_to_cart(self.ingredient1, 8),
-                       self.ingredient_to_cart(self.ingredient2ts, 4),
-                       self.ingredient_to_cart(self.ingredient2g, 1)]
-        self.assertEqual(ctx['ingredients'], ingredients)
+        ingredient_units = [self.model_to_cart(self.ingredient1pc, 8),
+                            self.model_to_cart(self.ingredient2ts, 4),
+                            self.model_to_cart(self.ingredient2g, 1)]
+        self.assertEqual(ctx['ingredient_units'], ingredient_units)
 
         # Edit cart
         self.client.post(reverse('cart'), {
@@ -136,15 +145,35 @@ class RecipesTestCase(TestCase):
         ctx = self.client.get(reverse('cart')).context
         recipes = [(self.recipe1, Decimal(1.5)), (self.recipe2, Decimal(1))]
         self.assertEqual(ctx['recipes'], recipes)
-        ingredients = [self.ingredient_to_cart(self.ingredient1, 6.5),
-                       self.ingredient_to_cart(self.ingredient2ts, 3),
-                       self.ingredient_to_cart(self.ingredient2g, 1)]
-        self.assertEqual(ctx['ingredients'], ingredients)
+        ingredient_units = [self.model_to_cart(self.ingredient1pc, 6.5),
+                            self.model_to_cart(self.ingredient2ts, 3),
+                            self.model_to_cart(self.ingredient2g, 1)]
+        self.assertEqual(ctx['ingredient_units'], ingredient_units)
+
+    def test_multiple_units(self):
+        self.client.get(reverse('add_to_cart', args=[self.recipe1.pk]))
+        self.client.get(reverse('add_to_cart', args=[self.recipe2.pk]))
+
+        self.ingredient2ts.factor = 10
+        self.ingredient2ts.save()
+        ctx = self.client.get(reverse('cart')).context
+        ingredient_units = [self.model_to_cart(self.ingredient1pc, 5),
+                            self.model_to_cart(self.ingredient2ts, 2),
+                            self.model_to_cart(self.ingredient2g, 1)]
+        self.assertEqual(ctx['ingredient_units'], ingredient_units)
+
+        self.ingredient2.primary_unit = self.g
+        self.ingredient2.save()
+        ctx = self.client.get(reverse('cart')).context
+        ingredient_units = [self.model_to_cart(self.ingredient1pc, 5),
+                            self.model_to_cart(self.ingredient2g, 21)]
+        self.assertEqual(ctx['ingredient_units'], ingredient_units)
 
     def test_localization(self):
         recipe = Recipe.objects.create(title='localization_test', recipe='')
         iir = IngredientInRecipe.objects.create(
-            recipe=recipe, ingredient=self.ingredient1, amount=1.5, order=1)
+            recipe=recipe, ingredient_unit=self.ingredient1pc, amount=1.5,
+            order=1)
         self.client.get(reverse('add_to_cart', args=[recipe.pk]))
         self.client.post(reverse('cart'), {
             'action': 'edit',
@@ -155,7 +184,7 @@ class RecipesTestCase(TestCase):
         # EN
         # On recipe page
         ctx = self.client.get(reverse('recipe', args=[recipe.pk])).context
-        self.assertEqual(list(ctx['ingredients']), [iir])
+        self.assertEqual(list(ctx['ingredient_units']), [iir])
         # On cart page
         ct = self.client.get(reverse('cart')).content.decode()
         self.assertTrue('0.5 localization_test' in ct)
@@ -165,7 +194,7 @@ class RecipesTestCase(TestCase):
         activate('nl')
         # On recipe page
         ctx = self.client.get(reverse('recipe', args=[recipe.pk])).context
-        self.assertEqual(list(ctx['ingredients']), [iir])
+        self.assertEqual(list(ctx['ingredient_units']), [iir])
         # On cart page
         ct = self.client.get(reverse('cart')).content.decode()
         self.assertTrue('0,5 localization_test' in ct)